SOLR-4555: When forceNew is used with CachingDirectoryFactory#get, the old CachValue should have it's path set to null.
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_4x@1454994 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/core/src/java/org/apache/solr/core/StandardDirectoryFactory.java
@@ -64,8 +64,10 @@ public class StandardDirectoryFactory extends CachingDirectoryFactory {
     if (val == null) {
       throw new IllegalArgumentException("Unknown directory " + dir);
     }
-    File dirFile = new File(val.path);
-    FileUtils.deleteDirectory(dirFile);
+    if (val.path != null) {
+      File dirFile = new File(val.path);
+      FileUtils.deleteDirectory(dirFile);
+    }
   }
 
   @Override
